ADD :: progressbar 모듈화 & 퍼센트 계산
